Modify: Search users feature. Add: email info
Updating the search users feature to work with the API's own request. Adding Alamofire request to SearchUser.swift. Adding an onChange method and lamda to ContentView.swift. Email info is needed in the API's own search, so email is also added to User.swift. Email now displayed to the list and it can be modified when editing a user and added when creating a new user.